modify augment.py in MAE & Botnet
